requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.3
cryptography==43.0.3
dynaconf==3.2.11
fastapi==0.136.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
jwcrypto==1.5.6
mosip-auth-sdk @ git+https://github.com/mosip/ida-auth-sdk.git@f322c59868123c2ef9410b5a3fecfd5dcdfd56af
numpy==2.2.6
opencv-python==4.12.0.88
packaging==26.2
pluggy==1.6.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pytest==9.0.3
requests==2.32.5
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.46.0
psycopg2-binary
python-dotenv